Update: add stop move, stop attack when champion death; add troop costs to loading game msg

+++ src/main/java/com/server/game/netty/sendObject/initialGameState/InitialPositionsSend.java
@@ -68,7 +68,7 @@ public class InitialPositionsSend implements TLVEncodable {
             dos.writeInt(burgHP);
 
             // Write number of champion positions
-            dos.writeShort(championPositionsData.size());
+            dos.writeShort((short) championPositionsData.size());
 
             // Write each champion's position data
             for (InitialPositionData positionData : championPositionsData) {
@@ -87,8 +87,7 @@ public class InitialPositionsSend implements TLVEncodable {
         return new AMatchBroadcastTarget(channel);
     }
     
-
-
+    
 
     //************* InitialPositionData class *************//
     @Data
